Move to financial data object module
