+++ core/domain/src/main/java/com/wansenai/bo/financial/PaymentDataExportBO.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.bo;
+package com.wansenai.bo.financial;
 
 import com.wansenai.utils.excel.ExcelExport;
 import lombok.AllArgsConstructor;
@@ -24,19 +24,25 @@ import java.math.BigDecimal;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class IncomeExpenseDataExportBO {
+public class PaymentDataExportBO {
 
-    @ExcelExport(value = "单据编号")
+    @ExcelExport(value = "供应商")
+    private String supplierName;
+
+    @ExcelExport(value = "付款单号")
     private String receiptNumber;
 
-    @ExcelExport(value = "往来单位/人员")
-    private String relatedPerson;
+    @ExcelExport(value = "采购单据编号")
+    private String purchaseReceiptNumber;
 
-    @ExcelExport(value = "收入/支出项目")
-    private String incomeExpenseName;
+    @ExcelExport(value = "应付欠款")
+    private BigDecimal paymentArrears;
 
-    @ExcelExport(value = "金额")
-    private BigDecimal incomeExpenseAmount;
+    @ExcelExport(value = "已付欠款")
+    private BigDecimal  prepaidArrears;
+
+    @ExcelExport(value = "本次付款")
+    private BigDecimal thisPaymentAmount;
 
     @ExcelExport(value = "备注")
     private String remark;
