+++ core/domain/src/main/java/com/wansenai/bo/financial/TransferAccountDataExportBO.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.bo;
+package com.wansenai.bo.financial;
 
 import com.wansenai.utils.excel.ExcelExport;
 import lombok.AllArgsConstructor;
